+++ pisync/util.py
